Combine the separated event publisher and listener into one
