+++ src/main/java/com/honeypot/common/event/ApplicationEventPublisher.java
@@ -1,14 +1,17 @@
 package com.honeypot.common.event;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Component;
 
 @Component
 @RequiredArgsConstructor
-public class ReactionCreatedEventPublisher {
+public class ApplicationEventPublisher {
 
-    private final ApplicationEventPublisher eventPublisher;
+    private final org.springframework.context.ApplicationEventPublisher eventPublisher;
+
+    public void publishEvent(CommentCreatedEvent event) {
+        eventPublisher.publishEvent(event);
+    }
 
     public void publishEvent(ReactionCreatedEvent event) {
         eventPublisher.publishEvent(event);
