--- src/main/java/com/honeypot/common/event/CommentCreatedEventPublisher.java
@@ -1,17 +1,0 @@
-package com.honeypot.common.event;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.ApplicationEventPublisher;
-import org.springframework.stereotype.Component;
-
-@Component
-@RequiredArgsConstructor
-public class CommentCreatedEventPublisher {
-
-    private final ApplicationEventPublisher eventPublisher;
-
-    public void publishEvent(CommentCreatedEvent event) {
-        eventPublisher.publishEvent(event);
-    }
-
-}
