--- src/main/java/com/honeypot/domain/notification/service/NotificationSendService.java
@@ -1,12 +1,0 @@
-package com.honeypot.domain.notification.service;
-
-import com.honeypot.common.event.CommentCreatedEvent;
-import com.honeypot.common.event.ReactionCreatedEvent;
-
-public interface NotificationSendService {
-
-    void send(CommentCreatedEvent commentCreatedEvent);
-
-    void send(ReactionCreatedEvent reactionCreatedEvent);
-
-}
